analyzer/requirements.txt
numpy
scipy
scikit-learn
grpcio
grpcio-tools
python-dotenv
umap-learn